lesson13: understanding state
State is a mechanism of React to update elements of html every time state object changes.

Creating component AddFishForm for the online store to input fish name, price, description etc. On submitting to the form we create a new object of fish by calling createFish.

At fish object creation and to the price, in order to have it in number use:
price: parseFloat(this.priceRef.value.value)
Please enter the commit message for your changes. Lines starting