lesson20: bi-directional data flow & live state editing
We have an input in the render function, where the value is set to our fish name and that in turn is living in state. Then when something changes on that input, we call func handleChange. React will then auto backspace that change, but we can get the value of what the person had hoped to type in event.currenttarget.value and we can take that value and update our fish by using event.currenttarget.value and we can only update the one field. Finaly we send all of the changes upstream to App component, because that's where our state lives with the updateFish function.