lesson23: component validation w/ PropTypes
Using PropTypes to declare the props we pass to the components and to know what to expect. Other options is to use the Facebook's utility called Flow. Another option is to extend JavaScript language with TypeScript to support types with variables. So to do PropTypes in Header coponent we use:

Header.propTypes = {
  tagline: PropTypes.string.isRequired
}

So to the App component going to use at the render function the <Header tagline="Fresh Seafood"> We use static propTypes = { }  in a class to have these PropTypes in every child component of this class.
Please enter the commit message for your changes. Lines starting